requirements.txt
flask==2.2.2
selenium==4.8.0
bs4==0.0.1
fpdf==1.7.2
flask_cors==3.0.10
requests==2.28.2
webdriver-manager
mysql-connector
-e .